refactor: extract the OIDC relying-party server into shared packages
Cire has been signing organisers in through OSN's OIDC provider since #302.
Pulse needs the same flow, and Musubi wants one shared sign-in across the
products — so the relying-party half moves out of cire/api and into the
shared workspace, where a second product can pick it up as a dependency
rather than a copy.

Four modules move, unchanged in behaviour:

  cire/api/src/lib/opaque-token.ts   -> @shared/crypto/tokens
  cire/api/src/services/oidc-login.* -> @shared/osn-auth-client/oidc-rp
  (new, split out of lib/cookie.ts)  -> @shared/osn-auth-client/cookie
  (new, from test-helpers)           -> @shared/osn-auth-client/testing/oidc-issuer

The cookie codec and the fake issuer were the only parts carrying cire's
name. Both now take what they need as arguments: the cookie builder takes
its name and flags, the issuer takes the return origin. Everything cire
still owns — three cookie names, its dev origin — stays in cire, as nine
one-line wrappers over the shared codec.

`OidcConfig` gains a required `txHmacInfo`. The transaction cookie is
MAC'd under a key derived from the OIDC client secret, and two products
sharing a secret would otherwise derive the same key: one product's
transaction cookie would verify at the other. Making it required rather
than defaulted means a new relying party cannot forget it. Cire's value
lives in the new lib/oidc.ts, imported by both the runtime config and the
test helper so the two cannot drift.

The 485-line RP test moves with the code, ported from bun:test to vitest
(shared/osn-auth-client's runner) and extended with one case the split
made worth asserting: a cookie minted under one product's `info` is
rejected under another's.

Cire's suite: 1492 pass, 0 fail. Full monorepo suite: 31/31 tasks green.
No route, cookie, or wire format changed.
