+++ shared/crypto/package.json
@@ -7,7 +7,8 @@
     ".": "./src/index.ts",
     "./jwk": "./src/jwk.ts",
     "./testing": "./src/testing.ts",
-    "./timing-safe": "./src/timing-safe.ts"
+    "./timing-safe": "./src/timing-safe.ts",
+    "./tokens": "./src/tokens.ts"
   },
   "scripts": {
     "check": "tsc --noEmit",
